Split most REST routes out into their own responsible classes.
