+++ includes/Core/Util/Reset.php
@@ -11,6 +11,11 @@
 namespace Google\Site_Kit\Core\Util;
 
 use Google\Site_Kit\Context;
+use Google\Site_Kit\Core\Permissions\Permissions;
+use Google\Site_Kit\Core\REST_API\REST_Route;
+use WP_REST_Server;
+use WP_REST_Request;
+use WP_REST_Response;
 
 /**
  * Class providing functions to reset the plugin.
@@ -43,6 +48,20 @@ final class Reset {
 	}
 
 	/**
+	 * Registers functionality through WordPress hooks.
+	 *
+	 * @since n.e.x.t
+	 */
+	public function register() {
+		add_filter(
+			'googlesitekit_rest_routes',
+			function( $routes ) {
+				return array_merge( $routes, $this->get_rest_routes() );
+			}
+		);
+	}
+
+	/**
 	 * Deletes options, user stored options, transients and clears object cache for stored options.
 	 *
 	 * @since 1.0.0
@@ -52,5 +71,34 @@ final class Reset {
 
 		// Call uninstaller.
 		require $this->context->path( 'uninstall.php' );
+	}
+
+	/**
+	 * Gets related REST routes.
+	 *
+	 * @since n.e.x.t
+	 *
+	 * @return array List of REST_Route objects.
+	 */
+	private function get_rest_routes() {
+		$can_setup = function() {
+			return current_user_can( Permissions::SETUP );
+		};
+
+		return array(
+			new REST_Route(
+				'core/site/data/reset',
+				array(
+					array(
+						'methods'             => WP_REST_Server::EDITABLE,
+						'callback'            => function( WP_REST_Request $request ) {
+							$this->all();
+							return new WP_REST_Response( true );
+						},
+						'permission_callback' => $can_setup,
+					),
+				)
+			),
+		);
 	}
 }
